(#794) Extract FFZ stuff from BttvFfz module to Ffz module
