+++ src/Bot/BttvFfz.hs
@@ -6,11 +6,8 @@
 {-# LANGUAGE LambdaCase #-}
 
 module Bot.BttvFfz
-  ( ffzCommand
-  , bttvCommand
-  , updateFfzEmotesCommand
+  ( bttvCommand
   , updateBttvEmotesCommand
-  , ffzUrlByName
   , bttvUrlByName
   ) where
 
@@ -18,11 +15,9 @@ import Bot.Replies
 import Control.Monad
 import Data.Aeson
 import Data.Aeson.Types
-import qualified Data.HashMap.Strict as HM
 import Data.List
 import qualified Data.Map as M
 import Data.Maybe
-import Data.Ord
 import Data.Proxy
 import qualified Data.Text as T
 import Effect
@@ -33,23 +28,6 @@ import Property
 import Reaction
 import Text.InterpolatedString.QM
 import Transport
-
-data FfzEmote = FfzEmote
-  { ffzName :: T.Text
-  , ffzLargestImageURL :: Maybe T.Text
-  }
-
-instance IsEntity FfzEmote where
-  nameOfEntity _ = "FfzEmote"
-  toProperties entity =
-    M.fromList $
-    catMaybes
-      [ return ("name", PropertyText $ ffzName entity)
-      , ("largestImageUrl", ) . PropertyText <$> ffzLargestImageURL entity
-      ]
-  fromProperties properties =
-    FfzEmote <$> extractProperty "name" properties <*>
-    pure (extractProperty "largestImageUrl" properties)
 
 data BttvEmote = BttvEmote
   { bttvName :: T.Text
@@ -85,62 +63,10 @@ instance FromJSON BttvEmote where
         (v .: "id")
   parseJSON invalid = typeMismatch "BttvEmote" invalid
 
-newtype FfzSet = FfzSet
-  { ffzSetEmotes :: [FfzEmote]
-  }
-
-instance FromJSON FfzSet where
-  parseJSON (Object v) = FfzSet <$> (v .: "emoticons")
-  parseJSON invalid = typeMismatch "FfzSet" invalid
-
-data FfzGlobalRes = FfzGlobalRes
-  { ffzGlobalResDefaultSets :: [Int]
-  , ffzGlobalResSets :: M.Map T.Text FfzSet
-  }
-
-instance FromJSON FfzGlobalRes where
-  parseJSON (Object v) = FfzGlobalRes <$> v .: "default_sets" <*> v .: "sets"
-  parseJSON invalid = typeMismatch "FfzGlobalRes" invalid
-
-newtype FfzRes = FfzRes
-  { ffzResEmotes :: [FfzEmote]
-  }
-
-instance FromJSON FfzEmote where
-  parseJSON (Object v) = FfzEmote <$> v .: "name" <*> (v .: "urls" >>= maxUrl)
-    where
-      maxUrl :: Value -> Parser (Maybe T.Text)
-      maxUrl (Object v') =
-        (\case
-           Nothing -> pure Nothing
-           (Just x) -> Just . ("https:" <>) <$> parseJSON x) =<<
-        pure ((`HM.lookup` v') =<< idx)
-        where
-          idx = listToMaybe $ sortOn Down $ HM.keys v'
-      maxUrl invalid = typeMismatch "FfzEmote" invalid
-  parseJSON invalid = typeMismatch "FfzEmote" invalid
-
-instance FromJSON FfzRes where
-  parseJSON (Object v) =
-    FfzRes <$> do
-      setId <- v .: "room" >>= (.: "set") :: Parser Int
-      v .: "sets" >>= (.: (T.pack $ show setId)) >>= (.: "emoticons")
-  parseJSON invalid = typeMismatch "FfzRes" invalid
-
-ffzUrl :: T.Text -> String
-ffzUrl channel = [qms|https://api.frankerfacez.com/v1/room/{encodedChannel}|]
-  where
-    encodedChannel = URI.encode $ T.unpack channel
-
 bttvUrl :: T.Text -> String
 bttvUrl channel = [qms|https://api.betterttv.net/2/channels/{encodedChannel}|]
   where
     encodedChannel = URI.encode $ T.unpack channel
-
-ffzCommand :: Reaction Message ()
-ffzCommand =
-  liftR (const $ selectEntities Proxy All) $
-  cmapR (T.concat . intersperse " " . map (ffzName . entityPayload)) sayMessage
 
 bttvCommand :: Reaction Message ()
 bttvCommand =
@@ -148,45 +74,6 @@ bttvCommand =
   cmapR (T.concat . intersperse " " . map (bttvName . entityPayload)) sayMessage
 
 
-updateFfzGlobalEmotes :: Reaction Message a
-updateFfzGlobalEmotes =
-  cmapR (const "https://api.frankerfacez.com/v1/set/global") $
-  jsonHttpRequestReaction $
-  cmapR
-    (\res ->
-       concatMap ffzSetEmotes $
-       mapMaybe ((`M.lookup` ffzGlobalResSets res) . T.pack . show) $
-       ffzGlobalResDefaultSets res) $
-  liftR (traverse $ createEntity Proxy) ignore
-
-cleanFfzCache :: Reaction Message a
-cleanFfzCache =
-  Reaction $ \_ -> void $ deleteEntities (Proxy :: Proxy FfzEmote) All
-
-updateFfzLocalEmotes :: Reaction Message a
-updateFfzLocalEmotes =
-  transR duplicate $
-  cmapR (twitchChannelName . senderChannel . messageSender) $
-  replyOnNothing "Works only in Twitch channels" $
-  cmapR ffzUrl $
-  jsonHttpRequestReaction $
-  cmapR ffzResEmotes $ liftR (traverse $ createEntity Proxy) ignore
-
-updateFfzEmotesCommand :: Reaction Message a
-updateFfzEmotesCommand = onlyForTwitch f
-  where
-    f =
-      cleanFfzCache <> updateFfzGlobalEmotes <> updateFfzLocalEmotes <>
-      Reaction (\msg -> replyMessage ("FFZ cache has been updated" <$ msg))
-
-ffzUrlByName :: T.Text -> Effect (Maybe String)
-ffzUrlByName name = do
-  emote <-
-    listToMaybe <$>
-    selectEntities
-      Proxy
-      (Filter (PropertyEquals "name" (PropertyText name)) All)
-  pure (T.unpack <$> (ffzLargestImageURL =<< (entityPayload <$> emote)))
 
 bttvUrlByName :: T.Text -> Effect (Maybe String)
 bttvUrlByName name = do
